pyproject.toml
[project]
name = "repl-nix-giftgenius"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.9",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "boto3>=1.35.39",
    "botocore>=1.35.39",
    "flask-dance>=7.1.0",
    "sqlalchemy>=2.0.35",
    "flask-cors>=5.0.0",
    "ebaysdk>=2.2.0",
    "etsy-python>=1.0.14",
    "cachetools",
    "etsy2>=0.7.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.149.0",
]
